refactor(daengle): 컨벤션에 맞게 변수명과 폴더명 수정
