+++ apps/daengle/src/api/api.ts
@@ -1,7 +1,7 @@
 import axios from 'axios';
 export const BASE_URL = process.env.NEXT_PUBLIC_BASE_URL;
 
-export const instance = axios.create({
+export const api = axios.create({
   baseURL: BASE_URL,
   headers: { 'Content-Type': 'application/json' },
 });
